DDCE-8619 move ConnectorSpecHelper to helpers package
