+++ test/uk/gov/hmrc/lisaapi/connectors/DesConnectorTestHelper.scala
@@ -17,7 +17,8 @@
 package uk.gov.hmrc.lisaapi.connectors
 
 import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.lisaapi.models._
+import uk.gov.hmrc.lisaapi.helpers.ConnectorSpecHelper
+import uk.gov.hmrc.lisaapi.models.*
 import uk.gov.hmrc.lisaapi.models.des.{DesFailure, DesResponse}
 
 import java.time.LocalDate
