+++ test/uk/gov/hmrc/lisaapi/helpers/ConnectorSpecHelper.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.lisaapi.connectors
+package uk.gov.hmrc.lisaapi.helpers
 
 import com.github.tomakehurst.wiremock.client.ResponseDefinitionBuilder
 import com.github.tomakehurst.wiremock.client.WireMock.*
